feat: disable open-webui-public - it seems to be interfering with open-webui
For some reason, logging into the internal (old) instance is loading up
the public instance and soon after it throws a 401. It looks like I'm
being routed to both instances for every request and it is not clear why
at the moment. I initially thought it was the volume mount but it looks
sufficiently distinct. I may need to look into the load balancer config
again.